edited public paths and documentation

+++ src/bitmap/pixel_data.rs
@@ -163,9 +163,8 @@ impl IndexMut<usize> for PixelData {
 impl std::fmt::Display for PixelData {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         for p in 0..self.pixels.len()
-        // 0..std::cmp::min(5, self.pixels.len()) as usize
         {
-            write!(f, "{}: {}\n", p, self.pixels[p as usize]).unwrap();
+            write!(f, "{}: {}\n", p, self.pixels[p]).unwrap();
         }
         write!(f, "")
     }
